finish save implementation
