+++ vesseval/state/__init__.py
@@ -2,7 +2,6 @@
 States enable a reactive implementation of widgets.
 """
 
-from .app import AppState, app_state
 from .contour import ContourState
 from .lib import (
     computed_state,
@@ -16,12 +15,10 @@ from .lib import (
     SequenceState,
     State,
 )
-from .image import DisplayImageState, ImageState, ResolutionState
+from .image import DisplayImageState, ImageState, ResolutionState, ImageConfigState
 from .point import PointState
 
 __all__ = [
-    "app_state",
-    "AppState",
     "ContourState",
     "computed_state",
     "FloatState",
@@ -35,6 +32,7 @@ __all__ = [
     "State",
     "DisplayImageState",
     "ImageState",
+    "ImageConfigState",
     "ResolutionState",
     "PointState",
 ]
