+++ vesseval/views/preprocessing/__init__.py
@@ -1,1 +1,1 @@
-from .main import PreprocessingView
+from .main import PreprocessingView, PreprocessingViewState
